added the compiled brixlib built by build.sh and commented out the individual scripts
Image tests work using the compiled file instead of all of the uncompiled files.
The map file is NOT working at this time, so debugging the compiled source using the uncompiled source is not working (at least in the chrome inspector. I believe it has to do with the path to the files during compilation.