quick fix: 981 unable to submit MC questions on Chrome Mac & Windows
After the Chrome update of 10/3/2013 it seems that descendant selectors are broken.
This caused no input elements to be selected to have changed event handlers set.

I've changed the selector for now MC so that it works. However that is a critical selector type which will probably affect other things unless it is fixed.