fix 918: MC issues after last Chrome update
It seems that the selector used to get the previous feedback elements to remove wasn't working because the properties of the parent weren't being taken into account.

I also noticed some extra descendant elements being selected because of a class that had been added to a div contained in the feedback, by the code that formats the feedback in the submitmanager. I removed that class.