minor cleanup-formatting & comments AND fix for chrome bug
This fix from the NeffOctDemo branch:
quick fix: 981 unable to submit MC questions on Chrome Mac & Windows

After the Chrome update of 10/3/2013 it seems element.querySelector is working as though the element was the document root and not considering the element itself or it's ancestors when evaluating a selector match.
This caused no input elements to be selected to have changed event handlers set.