define multiple choice response record type kept by the mc bric
Having this type will help implement the restoreState method because the response records will need to be restored.

Also fixed the code that was incorrectly referring to the score property when it should have used correctness. This had forced a fix by Leslie to get the correctAnswer key the hard way.